Monkey patch templates only when the panel is used
Fixes #6. This moves the monkey patching code from the AppConfig to the panel class, so that when the panel is first instantiated, it then does the monkey patching. This allows the app to be installed with the panel disabled without any overhead.
