--- template_profiler_panel/signals.py
@@ -1,3 +1,0 @@
-from django.dispatch import Signal
-
-template_rendered = Signal(providing_args=['instance', 'start', 'end', 'level'])
